Move JsonTransformer
